refactor: split utils/ into focused submodules, ref #7
- utils/mlflow_utils.py: MLflow helpers (setup_mlflow, load_run_params,
  get_checkpoint_path, parse_domains_param, load_checkpoint, etc.)
- utils/ckpt_utils.py: weight migration and state dict utilities
  (migrate_legacy_state_dict, strip_compiled_prefix, etc.)
- utils/run_loader.py: canonical reconstruct_from_run (split: str|list),
  reconstruct_model, config_from_runid; single AUTO_BLOCK_SIZE=512
- utils/__init__.py: public API with __all__; existing imports unchanged
- Remove orphaned functions (get_wte, get_ignored_tokens, get_best_ckpt,
  get_epoch_from_ckpt, get_top_counts, get_domain_configs_from_string)
- auc/compute_aucs.py: remove ~160 lines of duplicate helpers
- All callers updated to `from utils import ...`
