+++ train.py
@@ -1,16 +1,15 @@
 # %%
-import os, sys
+import os
+import sys
 from pathlib import Path
 import yaml
-from dataclasses import dataclass, asdict
+from dataclasses import asdict
 from pprint import pformat
-from dataclasses import fields as dc_fields
 import warnings
 import pandas as pd
 from auc.aucs import evaluate_aucs
 import torch
 from torch.utils.data import DataLoader
-from easydict import EasyDict
 import mlflow
 
 import logging
@@ -23,7 +22,6 @@ if ( DELPHI_DIR := Path(__file__).resolve().parent ) not in sys.path:
 from data.dataset import (
     DelphiDataset,
     DelphiCollateFn,
-    DelphiBatch,
     AgeSampler,
 )
 
@@ -44,13 +42,12 @@ from utils.trainer import (
 )
 
 from utils.cv_utils import get_data_partitions
-from utils.utils import load_domain_config, setup_mlflow
+from utils import load_domain_config, setup_mlflow, AUTO_BLOCK_SIZE # cache upper bound when block_size="auto"
 
 setup_mlflow()
 
 root_path = DELPHI_DIR / "data" / "transforms"
-ATTENTION_SCHEMES = yaml.safe_load( (DELPHI_DIR / "config" / "attention_schemes.yaml").read_text() )
-AUTO_BLOCK_SIZE = 512   # cache upper bound when block_size="auto"
+ATTENTION_SCHEMES = yaml.safe_load((DELPHI_DIR / "config" / "attention_schemes.yaml").read_text())
 
 torch.set_float32_matmul_precision("high")
 torch.backends.cudnn.allow_tf32 = True
@@ -188,9 +185,9 @@ def get_continuous_domains(domain_cfg):
 
 
 def get_dataloaders(
-    domain_cfg,
+    domain_cfg, 
     model,
-    test_fold,
+    test_fold, 
     block_size,
     batch_size,
     num_workers=4,
@@ -198,7 +195,6 @@ def get_dataloaders(
     no_event_insertion_mode="random",
     seed=42,
     subjects_include_list=None,
-    use_compile=True,
 ):
     train_ids, val_ids, test_ids = get_data_partitions(
         "./data/transforms/subject_lists", fold=test_fold
@@ -262,14 +258,6 @@ def get_dataloaders(
     eval_collate  = DelphiCollateFn(**collate_kwargs, training=False)
 
     loader_kwargs = dict(batch_size=batch_size, num_workers=num_workers, pin_memory=True)
-    if num_workers > 0 and use_compile:
-        # Fix: 
-        # Workers are re-forked at the end of each epoch (iterator GC). On exit,
-        # forked workers trigger LLVM thread cleanup which fails with pthread_join
-        # errors because torch.compile's LLVM thread pool is invalid in child processes.
-        # persistent_workers keeps workers alive across epochs so they never exit
-        # mid-training; the harmless LLVM error at program end is after training completes.
-        loader_kwargs["persistent_workers"] = True
 
     train_loader = DataLoader(train_dataset, shuffle=True,  collate_fn=train_collate, **loader_kwargs)
     valid_loader = DataLoader(valid_dataset, shuffle=False, collate_fn=eval_collate,  **loader_kwargs)
@@ -338,7 +326,6 @@ if __name__ == "__main__":
             no_event_insertion_mode=args.no_event_token_insertion_mode,
             seed=args.seed,
             subjects_include_list=args.subjects,
-            use_compile=not args.no_compile,
         )
 
         # ── Optimizer ─────────────────────────────────────────────────────
@@ -353,7 +340,7 @@ if __name__ == "__main__":
             warmup_iters   = args.warmup_iters,
             lr_decay_iters = args.lr_decay_iters,
         )
-        logging.info(f"Optimizer configuration: \n%s", pformat(asdict(optim_config), sort_dicts=False))
+        logging.info("Optimizer configuration: \n%s", pformat(asdict(optim_config), sort_dicts=False))
         
         optimizer, scheduler = configure_optimizers(model=model, cfg=optim_config, device_type=DEVICE)  
         
@@ -372,7 +359,7 @@ if __name__ == "__main__":
     else:
   
         ################################ FROM PREVIOUS RUN ################################
-        from utils.utils import config_from_runid 
+        from utils.run_loader import config_from_runid
         model, \
         dataloaders, \
         optim_config, optimizer_state, scheduler_state, \
